protocol080.md: Add file and improvements
This commit adds the file where protocol 0.80 will be developed
and includes the initial changes as to how client and server should behave.
Such as sending Position and Orientation Packets according to framerate
of the user. So when framerate is 60. Both packets should be send 60 times
a second as well.

And also includes how server should handle World Update when player is outside
of visible player range. That is 128 blocks. But should be around 132 for ping accounting.