protocol080.md: Combine input and orientation packets
Before people freak out why this was done.

As the protocol evolves and tries to reduce some of the main drawbacks with it.
Such as possibility of being in mismatch on client and server position sometimes even 2 blocks.

We currently have technically 2 inputs. Or well 3 considering gun but we do not count that here
as it doesnt have impact on physics.

2 packets with inputs have a massive drawback. That is they can arrive out of order to each other.
And even sending the ordered in Enet doesnt prevent this fully.

What can happen and HAPPENS in real life is that you can send input, orientation, input, orientation
and on client get orientation, input, input, orientation. Or other way.

Physics massively depend on orientation of the player. If you send move forward the server has to know
where forward is. Lets say you move your mouse really really fast and your orientation packets
(Some of them) get lost on the way. Happens. Now server doesnt know exactly where "forward" is.
And when you tell it to go forward it may move in direction that is out of match with clients one.

And client doesnt correct this as it already sent the correct orientation. So we have to rely on
position updates. But these have to be sent as much as possible. And 10 per second aint
cutting it here.

So to improve physics accuracy drastically lets merge keyboard inputs and orientation.

This way the server knows what the player wanted to do (Move forward. Or left. Or diagonally) and also knows
the correct orientation AT EXACTLY THAT MOMENT.