protocol.md: Introduce State Update
Instead of having position, orientation and input
packets we merge all of these together so we can
be sure the position we get is the position for
this exact input. Same applies for orientation
and also for the newly introduced velocity.

Velocity is being sent so the simulation which
now runs on both sides so client and server
can be identical or as close as possible to identical
on both sides. Velocity helps drastically in this.

We also send array of the last 30 inputs the player did
in each state update. (Packet loss protection)
And state data for the oldest input for the array of inputs.
0th element is newest. 29th oldest.

This way we can calculate where a player was precisely
up to 500ms in the past.

We also send state data for array of players to each player.

This is done so the animation client side can be smooth and the
actual position is on client side is precise.